api/requirements.txt
selenium 
python-dotenv